Change the syntax of the 'make' field.
Three forms are provided.
make = [ "ocp-build -init -scan" ]
can be rewritten as :

1.
make = [ [ "ocp-build" ; "-init" ; "-scan" ] ]
Note that in the futur, the ';' characters may disappear, this is currently temporary.

or 2.
make = [ # Sys.command "ocp-build -init -scan" # ]
For this case, at compilation time, some "cma" are added.
Note that in the futur, the '#' characters may disappear, this is currently temporary.

or 3.
make = [ " Sys.command \"ocp-build -init -scan\" " ]

Moreover, all these three forms can be mixed for a list of commands.