Handle optional dependencies.
Seems to work on small tests.

The idea is the following:

* you declare in your .opam some optional dependencies (using the same syntax as the 'depends' field)
* the optional packages are not used when solving the dependency constraints
* after the solver gave you a solution, we consider the universe of packages to be the one where optional dependencies become hard-dependencies. That means that we will automatically reinstall all the forward-dependencies of optional packages when it is installed.

Example:

* P5 depends on P1 and optionally depends on P2
* if we ask to install P5, P1 will be installed as well (but not P2)
* if later, we ask to install P2, then P5 will be reinstalled

Internally, we do a kind-of-ugly hack to propagate the optional dependencies through the solver, and we load twice the universe of packages (once without the optional dependencies, one with them). I guess we should wait for a proper support in the solver if we want to do better.