[solver optimization] New packages returned by the DOSE solver (not present in the request) could have an old version (among available versions). We introduce an optimization trying to always return the last version of newly computed packages.
Note that this remains an optimization as the solving problem being NP. For instance, the "opam upgrade" performed in #43 just after the "opam install" succeeds but could be unsuccessful in some other cases.
However, this optimization should be useful in practice as the user targets the last version in general.