Refactoring of the curl/rsync backend
Now the curl backend uses md5 checksums, which make it looks like very much the rsync backend, so this commit refactor a lot of common functionnalities.
As a side effect, it also fix the rule "if the archive is not there, then download by reading what's in url/", which make the behavior exactly the same for rsync/curl and opam-mk-repo (which incidentaly fix some bugs there which were present in the rsync backend there).

Note: the checksum support is not yet complete, as OPAM doesn't (yet) store checksums of external archives; but that's now quite easy to add.