Improve the reporting of errors
Now the sub-processes states are kept until all the processing is done, and the errors are displayed at the end.

This also help the error-recovery: in case of upgrade, we reinstall the old package; in case of reinstall we are a bit stuck and we remove all packages depending on this one. This solution is not yet perfect, as it may lead to some inconsistent global state. Hopefuly this will not happen very often ...

This commit fixes #130 and partially fixes #131