Add the ability to overwrite default variables.
It is now possible to tell opam to substitute variables in opam files in a more fine way. For instance, setting:

    $ opam config -var react:enable $ display "disable"
    $ opam install react
    $ opam config -var react:enable # display "enable"
    $ export foo_enabled=disable
    $ opam config -var react:enable # display "disable"

Note: all the 'opam config -var <var>' have a similar semantics to the variable %{var}% appearing in opam files, ie. they will be substituted by the same string when opam run.