Fix #470
Add an option to enable --autoremove when removing packages (by default it's off). You can also enable it by setting OPAMAUTOREMOVE.

The semantics is the following:

* `opam remove <pkg>` will simply remove the package
* `opam remove <pkg> -a` will remove the package and its dependencies
* `opam remove -a` will remove all the packages which have not been explicitely installed

(note: you can do `opam install <pkg>` to an already installed package to make it become a 'root' package which will not be uninstalled using `-a`).