Harder Better Safer 'opam update'
In OPAM 1.0, we were relying on each repository backend to tell us what the upgrade were. This was making the implementation of a new backend a little bit a pain (not it is really trivial) and was not very robust to occasional errors.

In the previous `master` version of OPAM, we were starting to compare the 'global' state (eg. what were in ~/.opam/opam, ~/.opam/desc ~/.opam/archives) with the 'repository' state (eg. what were in ~/.opam/repo/<repo>/packages/XXX/<package>). But we didn't have all the information we needed in the global state (for instance, we didn't know about urls, and additional files), the name of the files were changing, and so on.

So now, after a summer cleaning, and because now all the states ars stored globally, it is much easier and to compute the difference between the repository state and the global state, and to fix these properly and to tell the user. This is *much* more robust and simple than before.

Also, less caching (although we might add one back later if needed) so more deterministic results and simpler code.