Dev packages should live in the global state
Currently the temporary files for dev packages are stored in the repository states (~/.opam/repo/<repo>/tmp/<package>/), and this has caused much trouble (and make the code painfully complex). The the natural place for these file is the global state, under ~/.opam/packages.dev/<package>.

Now, all the metadata and data associated to a packages is stored in the global state; the only time we need to compute in which repo a package belongs to is when we want to download any associated `name.version+opam.tar.gz` archive: in this case we need to triger the right OpamRepository.pull_archive function.