Fixing more issues with packages removed from upstream
Hopefully it is now completely consistent: both upgrade and reinstall will run
additional checks, remove packages that are in the way (only when needed, or for
global upgrade) and trigger errors if the upstream removal makes the user
request impossible to fulfill.

There might be another, maybe simpler way to get a consistent result: see comments.