Allow more expressive package requests from the command-line
Closes #380 :
this allows to specify `pkgOPversion` and not just `pkg` or `pkg.version` as
before. Useful to manually request prerequisites for example, like `depA>=3` or
`depB<1.2`. This may also allow to get a better explanation when a prerequisite
is not available for some reason (when opam says `can't install A because B<1.2
is not available`, it's natural to try `opam install "B<1.2"` to understand why)

Does not yet allow more advanced requests, like logic formulas, though. This would
require deeper changes in the way we push requests through dose.

Also includes quite a bit of cleanup in OpamClient and the way we parse command-line
package arguments.