Turn complement of depopts into conflicts
This removes the current behaviour which turns depopts into hard deps depending on
the installed state of the packages before install, which is not always correct when
considering version constraints.

Instead, this turns the complement of the versions of packages appearing
in the depopts into conflicts, which is generally what is wanted.

NOTE however that the semantics of _conjunctions_ within depopts is unclear
yet, and those won't be considered specially yet.