Better default criteria
The previous criteria minimised all changes before minimising notuptodate, this
only minimises up and down-grades (i.e. all changes but new packages). This still
avoids the 'upgrading unrelated packages' problem, while avoiding the new problem
'not installing latest version because it has more dependances' (it manifests
itself with 'opam install ocp-build' which would give prehistoric 0.1 because
it didn't have any dependency).

Is this compatible with the currently deployed versions of aspcud though ?

Note: we'll still prefer the more specific '-notuptodate(requested),-changes' when
it becomes available, which avoids the mentionned problems but prioritises installing
the last versions of the packages you asked for.