A workaround older aspcuds
Ref: #1552

Can't say this solution makes me happy: I am checking if aspcud is a shell-script and
if it is, running it through 'bash -e' to make sure it fails on errors.

There really is no better solution that I can think of if we want to
stay compatible with this version, though and hopefully this hack can
be removed when we don't need to anymore. The 'empty solution'
safeguard is good, but far from enough because there is no way to tell
apart legitimate answer and failure when installing from an empty set
of packages (which does happen quite often in tests).