Fix update bug introduced in fef5f0704f59459dad119dbb30f63e92babc6d1f
OPAM would reinstall all packages after each update :(
The 'package states' are not well defined, and the interface between OpamRepository and OpamState
for that purpose is very fragile (it relies in the same lists being computed, in the same order,
from two modules). But since this will go in 1.3, there is no point in rewriting it now.