Add note on OpamVersionCompare (from dose) copyright
I've extracted it because it was the only dependence on dose3.debian, largely
independent from the rest of the Dose tools (Cudf uses ints for versions, so no
ordering issues), and mostly because we use it earlier than the dependency towards
Dose is otherwise needed (i.e. the opam-solver sub-library).