This PR will solve a fd leak
When `OpamProcess.command` is called with `allow_stdin:false` it will
create new channel for `stdin`, that will be duplicated in the child
process, but never closed in the parent.

The only place that currently disables stdin is in the `OpamCudf`
module.

Looks like, that this was the particular issue, that break BAP
installation ocaml/opam-repository#3994.

This issue starts to manifest on all our yosemite machines, when we're
trying to install BAP, after the aspcud was added to macports.