Tmp fix: cleanup after failed builds
with separated builds and install, this is no longer needed, but I was a bit too optimistic:
the repository is not there yet and lots of packages have a single  step including
install. When they fail during install, we still need this to not leave junk behind. This
should be reverted once all the packages in the repo have made the split (which we also
want for sandboxing).