Support parallel lib-ext compilation
Extends the parallel building available in lib-pkg to lib-ext by
declaring the relationships between the sub-projects and invoking them
directly. OCamlMakefile's architecture is inherently non-parallel, but
with care it is just possible to compile the native and bytecode
libraries simultaneously.

It is critical that make_deps=yes is hoisted into src_ext/Makefile or
both ncl and bcl may attempt to write ._d files simultaneously.

OCamlMakefile requires two fixes worthy of mention:
1. Compilation of .ml files which don't have a corresponding .mli file
   is problematic since both ocamlc and ocamlopt may attempt to write
   the .cmi file simultaneously. OCamlMakefile is heavily structured
   towards recursive make, so the best solution seems to be using locks
   to guard the compilation of .ml files. Added support for procmail's
   lockfile program - if this is not found, OCamlMakefile declares
   itself .NONPARALLEL (though for OPAM, this still allows the various
   libraries themselves to be compiled in parallel)
2. The rules for packs include a dependency on the .cmi file. I'm not
   sure if this is superfluous as presumably if the .mli file were
   declared, the .cmi file for the pack would be expressly generated as
   part of the build process, but its inclusion creates a
   bytecode/native code simultaneous write conflict. Since OPAM doesn't
   declare .mli headers for any of the packs, the dependency has simply
   been removed.

Various invocations of $(MAKE) require prefixing with + in order to
ensure that GNU make's jobserver mode (-jN) works.