Cache environment updates in a separate file.
The environment updates are cached in <root>/<switch>/environment
which is updated whenever the <root>/<switch>/state is updated.

This might be used by external tools (e.g. opam-manager) to read the
environment updates without having to depend on `OpamState`. This might
also permit to add a "costly topological sort" when computing the
environment updates.