package: allow reuse of build_package
Allow users to call build_package on a directory containing a package
that is ready to be built. It seems as if the the function should really
perform the functionality exposed through the `In_place variant,
delegating the work for extracting the package to the caller. It is left
like this to minimize the change necessary for the implementation of the
local sandbox feature.