Remove repos configuration files
repository configuration files were below the repository directory,
which was bad, and they now got cleaned up by the new synchronisation
functions that make sure no artifacts are left, which broke.

Besides, the only required and non-redundant information they contained
was the remote URL. This patch stores the remote urls of configured
repositories in a single file `ROOT/repo/repos-config`;

the selection of repositories is a list of names with the highest
priority repository first, and is stored in the global config file
`ROOT/config`. This removes the need for the "priority" integer field,
and is done with the idea that it can be moved to the switches sometime
soon to allow per-switch remotes. It's mostly an issue of UI now.

OTHER CHANGES:
- cleaner opam version ~/.opam upgrade code, added migration and some
  cleanup for 2.0~alpha
- changes of ~/.opam/config through a more specific function that
  requires the proper write lock, for more guarantees.
- Fixes #2538 (replace an iter by a fold)
- Put deprecated pin_option type aside (only left in
  OpamFile.Pinned_legacy)
- Internally, it's now possible to have remote-less repositories, which
  may be useful for some uses: local definitions can be stored there and
  will get used by opam without being touched by updates