Don't process reinstalls of parents on package upgrade
The idea was that you don't want to recompile `foo` twice on `opam
upgrade foo; opam upgrade`. But in practice, and after some discussions,
it seems this was more trouble than benefit, as you generally want the
minimum on `opam upgrade foo`, and esp. don't want to be forced to
recompile everything if `ocamlfind` has changed.