Add global `wrap-build/install/remove-commands:` fields
These are defined in ~/.opam/config and allow to wrap all command
invocation; this is primarily intended for testing purposes at the
moment, and the fields can only be set by hand.

The wrappers are resolved in the scope of the command that is
run (package variables are available). One intended usage is to restrict
read/write access of the commands on the testing frameworks (so that
e.g. a package that tries to install during its build phase will fail).