Correct multiple updates of a single environment variable
This ensures that if multiple packages update the same variable, `eval
$(opam update env)` remains stable (rather than appending again at each
call, and always returning `environment is not in sync`).

To do so, updates are first undone (with some fuzz) in reverse, then
reapplied in order. This assumes, at the moment, that the variable
was'nt modified in the meantime, though (undoing `FOO += "bar"` when
`FOO="baz:bar"` won't remove the `bar`) ; we could relax this and remove
the first occurence of `bar` when reverting `+= "bar"`.

This could also be applied to improve switching (revert variables from
the OLD switch, then apply updates for the new one): at the moment they
may linger.