New format upgrade: replace `available:` constraints with dependencies
following discussions, this new rewrite uses depends towards OCaml
rather than `available:` constraints on the `ocaml-version` and similar
variables;

this is much more pleasant and consistent, and allows e.g. correct
compiler upgrades. At the cost of automatic system compiler
selection (in the current state, you need to select the correct version
yourself).

Included opam change to not add implicit dependencies towards the
compiler packages anymore (the rewrite makes dependencies towards
`ocaml` explicit).

This is still in progress and needs testing.

--

The `ocaml` package is a wrapper packages that detects and defines the
`ocaml-native` and similar variables, using an OCaml script; it depends
on one of:
- `ocaml-base-compiler`, an official release of OCaml compiled by opam
- `ocaml-system`, a system compiler, coming from outside of opam (the
packages are just config scripts)
- `ocaml-variants`, all other or patched versions of OCaml

This allows us to map versions so that `depends: "ocaml" = "4.02.3"`
works correctly and allows `ocaml-variants.4.02.3+foo`.
All variants need to be in the same package, and described by the
version suffix, until we get an implementation of the `provides:` field,
to allow conflicts between all OCaml versions without too much trouble.