Travis: workaround while the repository is not rewritten
Travis basic tests rely on the standard repo to try and install 'lwt';
the 2.0 repo has not yet been updated to the new format, so this
fails because of a missing definition of 'ocaml-version' (references
to which will be removed by the 2.0~alpha2 rewrite).

This is a temporary workaround that manually sets this variable, in the meantime.