Better support for initialisation, separate OCaml-specific init
to a dedicated file, allow the file to set the default repositories.
This makes OCaml/opam.ocaml.org concerns clearly separate.

Define the `sys-ocaml-version` eval-variable by default, and add
`available:` constraints to system compilers generated by the repository
rewrite.

This is temporary, a dedicated file format (with repo names and urls,
the possibility to select a default initial switch...) is needed.