Unquote package names in opam files
(still allowing quoted ones)

This makes _some_ sense since we already have unquoted
`pkgname:variable`. However, in order to distinguish package names and
variables, we would need to add some syntax. A consistent way to do so,
syntax-wise, would be to prefix all variables with `:`, even when there
is no qualifying package (e.g. `build: [ :make "install" ]`).

The fact that it would be sound doesn't mean that it's what we want though.