Fix handling of unpacked source caches
move from `.opam-switch/packages.dev` to `.opam-switch/sources`, with
versions for unpinned packages, to avoid mixups when two different
versions or pinned/unpinned are needed concurrently (e.g. for
uninstalling).
Pinned package sources are versionless to avoid re-downloading if the
version is changed without a change in URL (the source is then cleaned
up only in case of URL change)