Don't always recompile src/client/opamGitVersion.ml
The conversion to OCamlMakefile lost an important subtlety of
shell/get-git-id.ml which is that it only actual writes to
src/client/opamGitVersion.ml if the file has changed. Unfortunately, the
Makefile rule touch'd the file regardless of what the script did, so
opam-client would always be recompiled, even if nothing had actually
changed.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>