Don't remove older-opam repository cache when in dev mode
Now that dev-version is generated reliably, it would be wiped at every
opam rebuild, which is generally unnecessary and takes a long time.

Of course, the downside is that, when using developer mode, you will get
the occasional segfault when modifying e.g. the OpamFile.OPAM.t type.
Just run `opam clean -r` when that happens.