New format for depexts
This closes #2919

This replaces the current obscure and verbose depexts format by lists of
system package names and filters based on opam variables, e.g. :

    depexts: [
      ["libssl-dev"] { os-distribution = "debian" | os-distribution = "ubuntu" }
      ["openssl-devel"] { os-distribution = "centos" & os-version >= "7.2" }
    ]

The old format still parses, and is rewritten by the package migration
function (older packages in 2.0 format won't be rewritten and may get
broken depexts, as we don't want to retain compatibility with beta
format versions)

All tags currently present in `opam-repository` are supported, except
for `source`, which is discarded (they were in the same namespace and
now need to be classified between `arch`, `os`, etc.)

Note that the rewrite needs to know about the variables defined and
their values. This is based on #3058 and
https://gist.github.com/AltGr/5bfc8cea6f01e74b95de79ceaba39369, version
4, but may still change.

Opam itself does not define these variables at the moment (except for
`arch` and `os`, but not with the same normalisation yet), and the
command-line format changes too, which means that current `opam-depext`
will be broken by the changes. The new format is

    opam list --external --vars arch=x86_64,os=linux,os-distribution=debian

instead of

    opam list --external=x86_64,linux,debian

which was more ambiguous and less generic, if shorter. The idea is that
the variables will later be built in opam, so the user won't need to
declare them for normal usage.