Add OpamConsole.clear_status
The translation of VT100 sequences to Windows API calls for the Console
mean that tricks relying on channel buffering no longer work (indeed, they
can't be implemented without a change in OCaml). This necessitates a
change in the way OpamConsole.status_line works, as it leaves the a
carriage_delete operation pending on the channel such that flushing the
channel will erase the status line.

OpamConsole.clear_status must now be called when OpamConsole.status_line
is no longer needed on-screen. It's not necessary to pair calls to
status_line and clear_status.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>