Add a field `subpath` in a package `url`.
The semantics of this fields is to restrict the files copied from the
directory `sources/<package>` to the directory `build/<package>`.

A later patch might allow the `rsync` backend to only synchronize in
`sources` the `subpath`. Another patch might allow a `VCS` backend not
to rebuild a pinned package when the 'sources' were modified, but only
outside of the 'subpath'.