Add opam-admin obsolescence check
Analysis for obsolete packages on a repository: a package A.V is
considered obsolete if there exists a version W, W > V, such that the
dependencies of A.W are equivalent, or less contraining than the
dependencies of A.V, and A.V has at least one dependent, and all its
dependents are also dependents of A.W.

Note that this is not enough to guarantee no packages become
uninstallable: a dependent could be shared but depend on a specific
combination of versions that is no longer available (e.g. A.1 & B.2 |
A.2 & B.1), and conflicts are not considered.