Don't guess the shell in OpamEnv.eval_string
It should always be possible to override the shell which opam reports
messages for. Add a parameter for the shell in OpamEnv.eval_string and
thread that resulting parameter through the libraries.

The install, pin, reinstall, remove and switch subcommands now all take
the --shell parameter as a result.

fixup