if 'ccache' is installed, add its cache directory to the sandbox (rw)
fixes #3395

This implementation ignores the CCACHE_DISABLED setting: it will add
the cache directory (rw) even if ccache is available but disabled in
the system. We could take this into account (there is a line of
the --print-config output that tells whether ccache is currently
disabled), but I thought it wasn't worth the trouble and extra script
complexity.