Fix CRLF processing of binary files in patches
The logic for the CRLF patch adaptation failed to take into account
mixed CRLF-ending files (i.e. binary files) which should obviously be
left alone.

read_lines updated to return a tri-state indicating if the file is
CRLF-encoded, LF-encoded or mixed. The patching logic is now altered to
accumulate patch lines and the patch is only adapted if both the target
file requires it and also the patch itself is consistent.