Add note on using `--locked` on already pinned packages
For technical reasons, that would be quite difficult to handle with the
current architecture, so be clear about it not being supported.

More detail:

we don't store the origin of the overlay opam file, in particular, we
don't know wether it came from `pin --edit`, a repo, the source, or the
source with `--locked` active. Hence it's impossible to determine
reliably wether the file has changed, and wether or not it should be
overwritten.

(e.g. running `opam install . --locked` then `opam install .` should
expose the same behaviour as running it the other way around, swapping
the two opam files. But we don't want `opam install .` to always
overwrite the overlay, in case that came from `pin --edit`)