upgrade: overlay opam file (#3528)
During the upgrade process, opam repo files are upgrades to 2.0, but not remaining switch opam files.
When pinning a package with opam1.2, an overlay opam file is stored internally. Then, when opam2.0.0 tries to install it, it takes it as is, and some check fail (the availability conditions error).
As the process of converting opam files from 1.2 to 2.0 is done at pin, repinning them with opam2.0 unblocks the situation.