Handle symlinks in bwrap
On some OSes, like CentOS and Arch Linux (so I've read), `/bin` is a symlink to `/usr/bin`, `/lib` is a symlink to `/usr/lib`, etc. 

Mounting them directly with `--ro-bind` resolves the symlinks. This breaks paths like `/bin/../libexec`: when `/bin` is a symlink to `/usr/bin`, `/bin/../libexec` points to `/usr/libexec`; but if it gets resolved first, it points to `/libexec`.

Fixes #3660