OpamParallel: expose the fact that the graph is unlabelled
I need to expose the fact that ActionGraph has unit-labelled edges, to
be able to write nice testing code using E.create (which requires
passing an edge label, and thus knowing the label type).

It doesn't make sense to constrain `OpamActionGraph` only to have unlabelled
edges (and it doesn't work as `OpamParralel.MakeGraph` returns a module
with abstract edge labels), because the assumption that edges are
unlabelled is currently made at the level of `OpamActionGraph`, as seen
in the API:

```ocaml
val iter:
  jobs:int ->
  command:(pred:(G.V.t * 'a) list -> G.V.t -> 'a OpamProcess.job) ->
  ?dry_run:bool ->
  ?mutually_exclusive:(G.V.t list list) ->
  G.t ->
  unit
```

If we wanted to support labelled edges, we would want to add the edge
label as an extra value in each element of the `pred` argument of the
`command` callback.

We discussed with Louis the fact that `OpamParallel` edges may become
labelled in the future. One may think that keeping the edge type
abstract makes this easier, but I would argue that revealing the type
is only making explicit an assumption that is made in the whole
implementation. If you move to labelled edge, you will have to update
the API, so revealing the edge type does not make it a "more breaking"
change.