Switch logging from Printf to Format
Using Format instead of Printf tidies up the Windows version of the
logging, since writing to a buffer and a channel become equivalent.
Technically, this now permits colour in log lines! It also opens the
door for using semantic tags instead of embedded escape sequences which
will allow the code to be further simplified.

The only downside is that the @ character now needs escaping in a few
messages.