Fix subtle bug with disappearing dependencies
Computation of dependencies was based on `OpamCudf.Graph.close_and_linearize`,
which is using a topological traversal of the dependency graph. Unfortunately,
when cycles appear (which is now allowed with `post` dependencies, but could
also happen before with malformed opam repos), this can cause dependencies to be
missed.

I believe this could manifest mysteriously in several bugs, but I dug it up on
an issue about `opam remove -a` no longer working, reporting a conflict while
attempting to remove `ocaml-config`.

Note 1: there are small changes in the API for simplicity's sake: the dependency
functions now return sets, and there is a separate sort function. We didn't use
the order anyway and always converted to sets, except in one `opam list` case
where we actually needed only the sort and not the dependency computing.

Note 2: I used a conservative approach (fix the bug first), but it would make
much sense to compute the cone directly from `OpamSolver` without going through
`OpamCudf` and back, which is quite coslty.