reftests: allow to pass env variables along
e.g. we had tests using `OPAMEXTERNALSOLVER`

We now use the CLI explicitely rather than pass to the environment (which we
were doing before, but is not a good idea in general because you don't want
tests to fail in ways where it's difficult to track back to the reason, here the
env)

However, it's complicated to make that go through dune. Since the tests with
this variable set already were disabled, it doesn't break anything at this
point. Adding an explicit error still.