Sandbox script: change handling of TMPDIR
The previous behaviour was to bind-mount $TMPDIR to /tmp within the sandbox, and
re-set TMPDIR accordingly ; however, this meant that absolute paths within
TMPDIR could be incompatible between the outside and the inside of the sandbox.

This became a visible problem with dune 2.8, which redefines TMPDIR to a
subdirectory (e.g. `/tmp/buildXXXX.dune`). If you happen to be running tests
with an OPAMROOT within $TMPDIR, your opam root will no longer be accessible,
and packages that run opam commands from within the sandbox (which is
discouraged, but well...) will break.

In our tests, run through dune, OPAMROOT gets created in
`/tmp/buildXXXX.dune/OPAM`, then packages get installed within a sandbox, this
is relocated to `/tmp/OPAM` but `OPAMROOT` and `OPAM_SWITCH_PREFIX` remained set
to below `/tmp/buildXXXX.dune`, and things broke.