Fix reverting additions to PATH-like variables
For environment variables which can be split (e.g. `PATH` on `:`),
OpamEnv.unzip_to attempts to find the point in the variable at which a
value was added. This is fine if a _single_ value was added, but fails
if the addition was multiple values (for example, if a setenv
instruction added two directories to `PATH` in one `+=`).

This is fixed by first splitting the value being searched according to
the same rule as the environmen variable and ensuring they all match in
order.