Add support for "opam switch -" (go to previous switch)
Use ordering of installed-switches instead of a new field

This avoids bumping the file version.

There are small differences in semantics though, as shown in the tests:
- after switching twice to the same switch, `opam switch -` now still gets you
  to the one before that
- it gets you to the 'last-known-good' switch, even if the last one was removed

I would argue that this new behaviour is ok, if not better.

Add a specific error in case the previous switch no longer exists

Co-authored-by: Louis Gesbert <louis.gesbert@ocamlpro.com>
Co-authored-by: R. Boujbel <rjbou@ocamlpro.com>