Solver: add support for requesting dependency-formulas
Using the same scheme as for enforcing the invariant (but with a distinct path,
we don't want the two to get confused in messages, etc.)

This is only in the internal API at the moment. The `request` type is extended
to allow a formula (instead of an atom list) in its `wish_install` field.