Reimplement --deps-only
Working around a set of bugs that were in the original version, but also
renouncing using the `--formula` code path because that was too limited for
using deps-only on multiple packages, and with unspecified versions.

This creates virtual packages at the client level that mimic the dependencies of
the specified packages (except for dependencies between them). This resolves a
few issues:

- these packages are always available (the original implem patched the packages
  but that could result in unavailable packages getting installed) ; they are
  however marked as avoid-version if unavailable as a cheap way to stay on the
  right version when possible.

- they exist in parallel to their "real" counterpart, so that these can still be
  installed if required through dependency chains, keeping consistency of the
  solution.

- they forbid different versions of their "real" counterpart, to ensure
  consistency.

- the `deps-of-xxx` package are removed from the solution, but could appear in
  conflict messages. I'd argue that it should make things understandable even if
  it is an implementation detail.