Make OpamConfigCommand.global_allowed_fields fully lazy
When it is not lazy, it causes other lazy values to be forced, notably
`OpamStd.Env.list`. This is unexpected, and a bit of a problem for people using
Opam as a library because they may want to set additional environment variables
before callin Opam.

Additionally, I'd expect this to be faster.