Fix invariant inference w.r.t. pinned packages
When inferring a switch invariant from base packages, opam filtered out
alternate versions of any package which is pinned. This causes all
pinned packages to appear to be unique versions and so they are then
inferred without constraint.

Pinned packages are no longer filtered when inferring the invariant -
only availability is used.