switch: remove compiler/base package
Since invariant is defined and updated, compiler_packages is no more
consistent the way it is. To keep field name, it is still named
`compiler-packages`, but its content is now invariant formula installed
packages and their recursive dependencies. It implies also that switch
state need to be update at each time an install/removal is done, or
invariant changed.

* Remove `universe.u_base` field
* Compute invariant packages instead of compiler packages
* Change PEF `base` field into `invariant-pkg`