Improve non UTF8 output
Currently when UTF8 is disabled via OPAMUTF8, `to_aligned_strings` assumes
`explicit` and also ignores the ascii symbol representation, example:

  - recompile carton-git               0.4.4   [uses logs]
  - recompile conduit-lwt-unix         5.1.0   [uses logs]

Only the '-' are displayed in yellow.
It's also a bit of a bummber that it has an extra field, so it doesn't play well
with awk & friends.

--

While the UTF8 will not ignore explicit and will output:

  <swirl symbol> carton-git               0.4.4   [uses logs]
  <swirl symbol> conduit-lwt-unix         5.1.0   [uses logs]

<swirl symbol> is in yellow

--

This diff changes the non-UTF8 to be:

  recompile carton-git               0.4.4   [uses logs]
  recompile conduit-lwt-unix         5.1.0   [uses logs]

'recompile' will be in yellow, and the non-UTF8 will have the same number of
fields as UTF8.

--

Also use a '*' instead of '--' when it's meant