Update switch environment file from env_hook
If the environment file is missing, recomputing it can take a noticeable
amount of time which delays every single return of the prompt when the
env_hook is enabled.

The use of atomic write on the environment file instead of a lock means
that the environment hooks can store the generated environment file if
it's missing, or needs regenerating.