Return the retrieved checksums on failure
Before this change it was not possible to get the actual checksum from
the API, because OPAM would download the file, compare checksums, emit
an error message and delete the file with the wrong checksum.

So to be able to get the actual checksum, one needed to disable the
validation and reimplement the validation on top of the unvalidated API
call. This PR returns the invalid checksums.