Re-apply eval-variables on upgrades
eval-variables are in opamrc from the 2.x de-OCaml-ing of opam, but
they're conceptually in the wrong place (they should be declared in the
repository where they can be synchronised and upgraded).

This is a problem for the new variables added in 2.1, because they don't
get added to the config of existing roots.

This alters the format upgrade to update pre-existing variables and add
any new ones.

Obviously, if the user had initialised with a different opamrc then
we're overwriting that, but that is the conceptual weakness of the field
being in opamrc.