Fix downloading URLs with invalid characters
OpamDownload.download downloads a given URL using the basename of the
URL as the filename. On Windows, where are there are many restrictions
on valid filenames, this causes a problem if the URL includes any query
string. Since OpamDownload passes the name used to the continuation, on
Windows the illegal characters are simply replaced with underscores.