Probe Git for Windows installations
In addition to not restarting the shell, it is possible that a user has
installed Git for Windows, but instructed its installer not to update
PATH. In this case, the installations will be referred to in the Windows
Registry. As a final fallback, opam init searches for these
installations and, if found, displays a warning suggesting that the user
may wish to reconfigure them (because then git will available to things
installed by opam, such as Dune), but also includes them in the list of
"gits" which can be manually selected with for git-location.