Don't trigger undefined variable warning for ?var
The implementation of the OpamSwitchState.get_dependencies_t displays
warnings on identifier resolution (which, with --strict enabled, are
fatal). This was incorrect, as this includes the resolution of variables
underneath the `?` operator (which tests for undefined variables).

The revised test relies on the fact that everything should have been
resolved except for `post` and displays errors after the filter has been
reduced.