GHA: Add OCaml 5.2.0 as one of the default tested compiler
Windows is an exception since MSVC isn't supported yet (will
be in 5.3.0), and Cygwin fails when used in combination with C++
stubs. So only MinGW x86_64 is tested because it's the only one that
both works and is reasonably fast to run (i686 is too slow since it's
using bytecode).