Cache shared archive sources without checkums
Without this change, sources that more than one pacakge share, but don't have
any specified checksum are not cached and hence downloaded once per package.
These can be downloaded only once per run instead of once per package,
this change enabled the per-run cache to make this happen.

For project like xapi, it means downloading a single tarball for 59
packages that are present in a single repository, every time it needs to
be built, saving time.

Closes #5638